==> analysis/screen/scripts/01_process_experiments.ipynb <==
## Process all counts for experiments to enrichment scores

import os
os.chdir('/media/data4/shared/ipsc_manuscript/')

run analysis/screen/input/ScreenProcessing/process_experiments.py -h

## Important package versions

import pandas as pd
print(pd.__version__)

import numpy as np
np.version.version

# Differentiation

## Coding

run analysis/screen/input/ScreenProcessing/process_experiments.py \
analysis/screen/output/01_process_experiments/config/Differentiation_Coding.txt \
analysis/screen/input/ScreenProcessing/library_tables \
--plot_extension "off"

## lncRNA: CRiNCL

run analysis/screen/input/ScreenProcessing/process_experiments.py \
analysis/screen/output/01_process_experiments/config/Differentiation_CRiNCL.txt \
analysis/screen/input/ScreenProcessing/library_tables \
--plot_extension "off"

## lncRNA: Common

run analysis/screen/input/ScreenProcessing/process_experiments.py \
analysis/screen/output/01_process_experiments/config/Differentiation_Common.txt \
analysis/screen/input/ScreenProcessing/library_tables \
--plot_extension "off"

# Proliferation

## Coding

run analysis/screen/input/ScreenProcessing/process_experiments.py \
analysis/screen/output/01_process_experiments/config/Proliferation_Coding.txt \
analysis/screen/input/ScreenProcessing/library_tables \
--plot_extension "off"

## lncRNA: CRiNCL

run analysis/screen/input/ScreenProcessing/process_experiments.py \
analysis/screen/output/01_process_experiments/config/Proliferation_CRiNCL.txt \
analysis/screen/input/ScreenProcessing/library_tables \
--plot_extension "off"

## lncRNA: Common

run analysis/screen/input/ScreenProcessing/process_experiments.py \
analysis/screen/output/01_process_experiments/config/Proliferation_Common.txt \
analysis/screen/input/ScreenProcessing/library_tables \
--plot_extension "off"

run analysis/screen/input/ScreenProcessing/process_experiments.py \
analysis/screen/output/01_process_experiments/config/Proliferation_Common_validation.txt \
analysis/screen/input/ScreenProcessing/library_tables \
--plot_extension "off"
